[#21] refactor: 파일 경로 수정
