--- src/components/types/search.ts
@@ -1,13 +1,0 @@
-export interface MovieResponseList {
-  docId: string;
-  genre: string;
-  nation: string;
-  poster: string;
-  prodYear: string;
-  title: string;
-}
-
-export interface searchMovieBar {
-  message: string;
-  data: MovieResponseList[];
-}
